Bug 358408 - Patch proposal for build warnings removal

+++ services/breakpoints.c
@@ -435,7 +435,8 @@ int check_breakpoints_on_memory_read(Context * ctx, ContextAddress address, void
             ContextAddress mem_base = 0;
             ContextAddress mem_size = 0;
             if (context_get_canonical_addr(ctx, address, &mem, &mem_addr, &mem_base, &mem_size) < 0) return -1;
-            if (mem_base + mem_size - mem_addr < sz) sz = mem_base + mem_size - mem_addr;
+            /* XXX : fle : check if cast is not an error */
+            if ((size_t) (mem_base + mem_size - mem_addr) < sz) sz = (size_t) (mem_base + mem_size - mem_addr);
             while (l != &instructions) {
                 BreakInstruction * bi = link_all2bi(l);
                 size_t i;
@@ -470,7 +471,8 @@ int check_breakpoints_on_memory_write(Context * ctx, ContextAddress address, voi
             ContextAddress mem_base = 0;
             ContextAddress mem_size = 0;
             if (context_get_canonical_addr(ctx, address, &mem, &mem_addr, &mem_base, &mem_size) < 0) return -1;
-            if (mem_base + mem_size - mem_addr < sz) sz = mem_base + mem_size - mem_addr;
+            /* XXX : fle : check if cast is not an error */
+            if ((size_t) (mem_base + mem_size - mem_addr < sz)) sz = (size_t) (mem_base + mem_size - mem_addr);
             while (l != &instructions) {
                 BreakInstruction * bi = link_all2bi(l);
                 l = l->next;
@@ -2379,14 +2381,16 @@ static void event_code_unmapped(Context * ctx, ContextAddress addr, ContextAddre
      */
     int cnt = 0;
     while (size > 0) {
-        size_t sz = size;
+        /* XXX : fle : check if cast is not an error */
+        size_t sz = (size_t) size;
         LINK * l = instructions.next;
         Context * mem = NULL;
         ContextAddress mem_addr = 0;
         ContextAddress mem_base = 0;
         ContextAddress mem_size = 0;
         if (context_get_canonical_addr(ctx, addr, &mem, &mem_addr, &mem_base, &mem_size) < 0) break;
-        if (mem_base + mem_size - mem_addr < sz) sz = mem_base + mem_size - mem_addr;
+        /* XXX : fle : check if cast is not an error */
+        if ((size_t) (mem_base + mem_size - mem_addr) < sz) sz = (size_t) (mem_base + mem_size - mem_addr);
         while (l != &instructions) {
             int i;
             BreakInstruction * bi = link_all2bi(l);
